+++ src/dolfin_mech/core/__init__.py
@@ -4,5 +4,6 @@ from .compute_error import compute_error
 from .FOI import FOI
 from .QOI import QOI
 from .TimeVaryingConstant import TimeVaryingConstant
+from .XDMFFile import XDMFFile
 
-__all__ = ["FOI", "QOI", "compute_error", "TimeVaryingConstant"]
+__all__ = ["FOI", "QOI", "compute_error", "TimeVaryingConstant", "XDMFFile"]
